Fix spinlock tests

+++ tests/test-spinlock.cpp
@@ -135,7 +135,7 @@ TEST_CASE("spinlock.c", "[spinlock]") {
             REQUIRE(spinlock_try_lock(&lock));
 #if DEBUG == 1
             long thread_id = syscall(__NR_gettid);
-            REQUIRE(lock.lock == (uint8_t)thread_id);
+            REQUIRE(lock.lock == (uint32_t)thread_id);
 #else
             REQUIRE(lock.lock == SPINLOCK_LOCKED);
 #endif
@@ -178,7 +178,7 @@ TEST_CASE("spinlock.c", "[spinlock]") {
 
 #if DEBUG == 1
             long thread_id = syscall(__NR_gettid);
-            REQUIRE(lock.lock == (uint8_t)thread_id);
+            REQUIRE(lock.lock == (uint32_t)thread_id);
 #else
             REQUIRE(lock.lock == SPINLOCK_LOCKED);
 #endif
@@ -194,7 +194,7 @@ TEST_CASE("spinlock.c", "[spinlock]") {
 
 #if DEBUG == 1
             long thread_id = syscall(__NR_gettid);
-            REQUIRE(lock.lock == (uint8_t)thread_id);
+            REQUIRE(lock.lock == (uint32_t)thread_id);
 #else
             REQUIRE(lock.lock == SPINLOCK_LOCKED);
 #endif
@@ -218,7 +218,7 @@ TEST_CASE("spinlock.c", "[spinlock]") {
 
 #if DEBUG == 1
             long thread_id = syscall(__NR_gettid);
-            REQUIRE(lock.lock == (uint8_t)thread_id);
+            REQUIRE(lock.lock == (uint32_t)thread_id);
 #else
             REQUIRE(lock.lock == SPINLOCK_LOCKED);
 #endif
